requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.4
pydantic-settings==2.2.1
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
PyJWT==2.8.0
itsdangerous==2.1.2
python-multipart==0.0.9
